+++ src/toolpath_exporter/toolpath-exporter-fcode.h
@@ -134,50 +134,11 @@ class ToolpathExporterFcode : public QObject {
   ToolpathExporterFcode(QTransform move_translate,
                         int dpi,
                         const QJsonObject* param,
-                        const QString* thumbnail) noexcept
-      : move_translate_(move_translate) {
-    qInfo() << "ToolpathExporterFcode init";
-    parseParam(param);
-    updateMovetoPipeline();
-    setDpi(dpi);
-
-    if (is_v2_) {
-      if (is_rotary_task_ || with_custom_origin_) {
-        magic_number_ = 4;
-      } else {
-        magic_number_ = 3;
-      }
-    } else {
-      magic_number_ = 1;
-    }
-
-    QString type = param->value("type").toString();
-    if (type == "gcode") {
-      is_gcode_ = true;
-      gen = std::make_shared<FCodeGeneratorG>();
-    } else if (is_v2_) {
-      gen = std::make_shared<FCodeGeneratorV2>(thumbnail, magic_number_);
-    } else {
-      gen = std::make_shared<FCodeGeneratorV1>(thumbnail);
-    }
-    gen_ = gen.get();
-    gen_->add_metadata("START_WITH_HOME", with_custom_origin_ ? "0" : "1");
-    gen_->add_metadata("3D_CURVE_TASK", is_3d_task_ ? "1" : "0");
-
-    gen_->set_time_est_z_speed(config_.z_speed);
-    setTravelSpeed(config_.travel_speed);
-  }
-
-  std::string toString() { return gen_->to_string(); };
-
-  void save(QDataStream* out) {
-    out->writeRawData(toString().c_str(), gen_->total_length());
-  }
-
-  float getTimeCost() { return gen_->get_time_cost(); }
-
-  QJsonObject getMetadata() { return gen_->get_metadata(); }
-
+                        const QString* thumbnail) noexcept;
+  std::string toString();
+  void save(QDataStream* out);
+  float getTimeCost();
+  QJsonObject getMetadata();
   bool convertStack(const QList<LayerPtr>& layers,
                     QProgressDialog* dialog = nullptr);
 
@@ -188,220 +149,12 @@ public Q_SLOTS:
   void handleCancel();
 
  private:
-  void parseParam(const QJsonObject* paramPtr) {
-    QJsonObject param = *paramPtr;
-    if (param.contains("job_origin")) {
-      with_custom_origin_ = true;
-      config_.job_origin = QPointF(param["job_origin"].toArray()[0].toDouble(),
-                                   param["job_origin"].toArray()[1].toDouble());
-    }
-    float spinning_axis_coord = param["spin"].toDouble(-1);
-    if (spinning_axis_coord >= 0) {
-      is_rotary_task_ = true;
-      config_.spinning_axis_coord = spinning_axis_coord / canvas_mm_ratio - config_.job_origin.y();
-      rotary_y_ratio_ = param["rotary_y_ratio"].toDouble(1);
-    }
-
-    QJsonObject workarea = param["workarea"].toObject();
-    int width = workarea["width"].toInt();
-    int height = workarea["height"].toInt();
-    QString model = param["model"].toString();
-    float default_path_travel_speed = 7500;
-    QJsonObject default_path_acc = {};
-    if (model == "fbm1") {
-      hardware_ = HardwareType::beamo;
-      config_.fg_pwm_limit = 1500;
-    } else if (model == "fbb1p") {
-      hardware_ = HardwareType::BeamboxPro;
-      config_.fg_pwm_limit = 1500;
-    } else if (model == "fhexa1") {
-      hardware_ = HardwareType::HEXA;
-      config_.support_rel_z_move = true;
-    } else if (model == "ado1") {
-      hardware_ = HardwareType::Ador;
-      is_v2_ = true;
-      config_.support_modules = true;
-      config_.support_rel_z_move = true;
-      config_.support_rotary_z_motion = true;
-      default_path_travel_speed = 3600;
-      default_path_acc["x"] = 500;
-      default_path_acc["y"] = 500;
-      if (param.contains("prespray")) {
-        QJsonArray prespray_arr = param["prespray"].toArray();
-        config_.prespray =
-            QRectF(prespray_arr[0].toDouble(), prespray_arr[1].toDouble(),
-                   prespray_arr[2].toDouble(), prespray_arr[3].toDouble());
-      }
-    } else if (model == "fbb2") {
-      hardware_ = HardwareType::BB2;
-      is_v2_ = true;
-      config_.z_speed = 5.16;
-      config_.support_rel_z_move = true;
-      config_.z_premove_speed = 140;
-      config_.z_premove_x = 0.0127;
-      config_.z_premove_y = 0.0064;
-      config_.z_premove_z = 0.0005;
-      default_path_acc["x"] = 1000;
-      default_path_acc["y"] = 1000;
-    } else if (model.startsWith("fhx2rf")) {
-      hardware_ = HardwareType::RF;
-      is_v2_ = true;
-    } else {
-      // default beambox
-      hardware_ = HardwareType::Beambox;
-      config_.fg_pwm_limit = 1500;
-    }
-    work_area_mm_ = QSizeF(width, height);
-    config_.dpmm_preview = 500.0 / width;
-
-    if (config_.support_modules) {
-      QJsonObject offset_dict = param["mof"].toObject();
-      for (QString module_key : offset_dict.keys()) {
-        QJsonArray offset = offset_dict[module_key].toArray();
-        config_.module_offsets[module_key.toInt()] =
-            QPointF(offset[0].toDouble(), offset[1].toDouble());
-      }
-    }
-
-    if (param.contains("diode")) {
-      config_.enable_diode = true;
-      QJsonArray diode_offset = param["diode"].toArray();
-      config_.diode_offset =
-          QPointF(diode_offset[0].toDouble(), diode_offset[1].toDouble());
-    }
-    config_.enable_autofocus = param["af"].toBool();
-    config_.enable_custom_backlash = param["cbl"].toBool();
-    config_.enable_fast_gradient = param["fg"].toBool();
-    config_.enable_mock_fast_gradient = param["mfg"].toBool();
-    config_.enable_pwm = !param["no_pwm"].toBool();
-    config_.enable_multipass_compensation = param["mpc"].toBool();
-    config_.enable_segmentation = param["segment"].toBool(true);
-    config_.enable_rotary_z_move = param["rotary_z_motion"].toBool() && config_.support_rotary_z_motion;
-    config_.is_one_way_printing = param["owp"].toBool();
-    config_.is_diode_one_way_engraving = param["diode_owe"].toBool();
-    config_.is_reverse_engraving = param["rev"].toBool();
-    config_.min_speed = param["min_speed"].toDouble(3);
-    config_.travel_speed = param["ts"].toDouble(7500);
-    config_.a_travel_speed = param["ats"].toDouble(2000);
-    config_.path_travel_speed = param["pts"].toDouble(default_path_travel_speed);
-    config_.vector_speed_constraint = param["vsl"].toDouble(0);
-    config_.curve_speed_constraint = param["csl"].toDouble(0) / 60;
-    config_.padding_acc = param["acc"].toDouble(4000);
-    config_.min_engraving_padding = param["mep"].toDouble(NAN);
-    config_.min_printing_padding = param["mpp"].toDouble(NAN);
-    config_.z_offset = param["z_offset"].toDouble(0);
-    config_.loop_compensation = param["loop_compensation"].toDouble() / canvas_mm_ratio;
-    config_.printing_top_padding = param["ptp"].toInt(10);
-    config_.printing_bot_padding = param["pbp"].toInt(10);
-    if (param.contains("nv")) {
-      nozzle_settings.voltage = param["nv"].toDouble();
-    }
-    if (param.contains("npw")) {
-      nozzle_settings.pulse_width = param["npw"].toDouble();
-    }
-    if (param.contains("acc_override")) {
-      config_.fill_acc = param["acc_override"].toObject()["fill"].toObject();
-      config_.path_acc = param["acc_override"].toObject()["path"].toObject();
-    } else {
-      config_.path_acc = default_path_acc;
-    }
-
-    QJsonArray clip = param["mask"].toArray();
-    if (clip.size() == 4) {
-      for (int i = 0; i < 4; i++) {
-        config_.workarea_clip[i] = clip[i].toDouble();
-      }
-    }
-
-    QJsonObject curve_obj = param["curve_engraving"].toObject();
-    if (!curve_obj.isEmpty()) {
-      qInfo() << "Set curve engraving data";
-      QJsonObject bbox = curve_obj["bbox"].toObject();
-      float box_left = bbox["x"].toDouble();
-      float box_top = bbox["y"].toDouble();
-      if (with_custom_origin_) {
-        box_left -= config_.job_origin.x();
-        box_top -= config_.job_origin.y();
-      }
-      config_.workarea_clip[3] = qMax(config_.workarea_clip[3], box_left);
-      config_.workarea_clip[0] = qMax(config_.workarea_clip[0], box_top);
-      float box_width = bbox["width"].toDouble();
-      float box_right = width - box_left - box_width;
-      config_.workarea_clip[1] = qMax(config_.workarea_clip[1], box_right);
-      float box_height = bbox["height"].toDouble();
-      float box_bottom = height - box_top - box_height;
-      config_.workarea_clip[2] = qMax(config_.workarea_clip[2], box_bottom);
-
-      QJsonArray points = curve_obj["points"].toArray();
-      int point_size = points.size();
-      if (point_size >= 3) {
-        is_3d_task_ = true;
-        // add 0.01 to avoid clipping the boundary
-        float space = 0.01;
-        curve_settings.bbox = QRectF(box_left - space, box_top - space, box_width + 2 * space, box_height + 2 * space);
-        curve_settings.interpolator.set_bounding_box(box_left, box_top, box_right, box_bottom);
-        QJsonArray gap = curve_obj["gap"].toArray();
-        curve_settings.gap = QPointF(gap[0].toDouble(), gap[1].toDouble());
-
-        for (int i = 0; i < point_size; i++) {
-          QJsonArray point = points[i].toArray();
-          float x = point[0].toDouble();
-          float y = point[1].toDouble();
-          float z = point[2].toDouble();
-          if (with_custom_origin_) {
-            x -= config_.job_origin.x();
-            y -= config_.job_origin.y();
-          }
-          curve_settings.interpolator.add_point(x, y, z);
-        }
-        curve_settings.interpolator.setup();
-        curve_settings.safe_height = curve_obj["safe_height"].toDouble(NAN);
-      }
-    }
-  }
-
+  void parseParam(const QJsonObject* paramPtr);
   void setDpi(int dpi) {
-    qInfo() << "Set dpi" << dpi;
-    if (dpi < 200) {
-      // low: 127
-      config_.dpmm_x = config_.dpmm_y = 5;
-      config_.print_modes[0] = 'P', config_.print_modes[1] = 'L';
-    } else if (dpi < 500) {
-      // medium: 254
-      config_.dpmm_x = config_.dpmm_y = 10;
-      config_.print_modes[0] = 'Q', config_.print_modes[1] = 'M';
-    } else if (dpi < 1000) {
-      // high: 508
-      config_.dpmm_x = config_.dpmm_y = 20;
-      config_.print_modes[0] = 'R', config_.print_modes[1] = 'H';
-    } else {
-      // ultra: 1016, only increase dpmm in y direction
-      config_.dpmm_x = 20, config_.dpmm_y = 50;
-      config_.print_modes[0] = 'R', config_.print_modes[1] = 'H';
-    }
-
-    // Calculate length in y px
-    path_utils_.setLengthRatio(config_.dpmm_x / config_.dpmm_y, 1);
-    path_utils_.setLoopCompensation(config_.loop_compensation * config_.dpmm_y);
-
-    transform_laser_ = QTransform::fromScale(config_.dpmm_x / canvas_mm_ratio,
-                                             config_.dpmm_y / canvas_mm_ratio);
-    transform_printing_ =
-        QTransform::fromScale(config_.dpmm_printing / canvas_mm_ratio,
-                              config_.dpmm_printing / canvas_mm_ratio);
-    transform_preview_laser_ =
-        QTransform::fromScale(config_.dpmm_preview / config_.dpmm_x,
-                              config_.dpmm_preview / config_.dpmm_y);
-    transform_preview_printing_ =
-        QTransform::fromScale(config_.dpmm_preview / config_.dpmm_printing,
-                              config_.dpmm_preview / config_.dpmm_printing);
+    // deprecated
   }
 
-  void setTransform(QTransform transform = QTransform()) {
-    global_transform_ =
-        transform * move_translate_ *
-        (is_printing_layer_ ? transform_printing_ : transform_laser_);
-  }
+  void setTransform(QTransform transform = QTransform());
   QTransform getPreviewTransform() {
     return is_printing_layer_ ? transform_preview_printing_
                               : transform_preview_laser_;
